requirements.txt
torch==2.3.0+cu126 --index-url https://download.pytorch.org/whl/cu126
torchvision>=0.15.0
opencv-python>=4.8.0
mediapipe>=0.10.0
customtkinter>=5.2.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
